Transpile addons.js with Babel
