+++ lib/core/src/server/common/babel-loader.js
@@ -1,0 +1,13 @@
+import { includePaths, excludePaths } from '../config/utils';
+
+export default options => ({
+  test: /\.(mjs|jsx?)$/,
+  use: [
+    {
+      loader: 'babel-loader',
+      options,
+    },
+  ],
+  include: includePaths,
+  exclude: excludePaths,
+});
