+++ lib/core/src/server/common/common-preset.js
@@ -1,0 +1,11 @@
+import loadCustomBabelConfig from '../utils/load-custom-babel-config';
+
+import babelConfig from './babel';
+
+export const babel = async (_, options) => {
+  const { configDir, presets } = options;
+
+  return loadCustomBabelConfig(configDir, () =>
+    presets.apply('babelDefault', babelConfig(options), options)
+  );
+};
