+++ lib/core/src/server/dev-server.js
@@ -47,7 +47,9 @@ export default async function(options) {
           },
           (err, stats) => {
             managerTotalTime = process.hrtime(startTime);
-            if (err || stats.hasErrors()) {
+            if (err) {
+              reject(err);
+            } else if (stats.hasErrors()) {
               reject(stats);
             } else {
               resolve(stats);
