+++ lib/core/src/server/manager/manager-config.js
@@ -2,9 +2,10 @@ import loadPresets from '../presets';
 import loadCustomPresets from '../common/custom-presets';
 
 async function getManagerWebpackConfig(options, presets) {
+  const babelOptions = await presets.apply('babel', {}, options);
   const entries = await presets.apply('managerEntries', [], options);
 
-  return presets.apply('managerWebpack', {}, { ...options, entries });
+  return presets.apply('managerWebpack', {}, { ...options, babelOptions, entries });
 }
 
 export default async options => {
