+++ lib/core/src/server/manager/manager-preset.js
@@ -23,3 +23,5 @@ export async function managerEntries(_, options) {
 export async function addons(_, options) {
   return loadCustomAddons(options);
 }
+
+export * from '../common/common-preset';
