+++ lib/core/src/server/manager/manager-webpack.config.js
@@ -10,12 +10,13 @@ import findCacheDir from 'find-cache-dir';
 import { version } from '../../../package.json';
 import { getManagerHeadHtml } from '../utils/template';
 import { loadEnv } from '../config/utils';
+import babelLoader from '../common/babel-loader';
 
 const coreDirName = path.dirname(require.resolve('@storybook/core/package.json'));
 const context = path.join(coreDirName, '../../node_modules');
 const cacheDir = findCacheDir({ name: 'storybook' });
 
-export default ({ configDir, configType, entries, outputDir, cache }) => {
+export default ({ configDir, configType, entries, outputDir, cache, babelOptions }) => {
   const { raw, stringified } = loadEnv();
   const isProd = configType === 'PRODUCTION';
 
@@ -55,6 +56,9 @@ export default ({ configDir, configType, entries, outputDir, cache }) => {
       new CaseSensitivePathsPlugin(),
       new Dotenv({ silent: true }),
     ],
+    module: {
+      rules: [babelLoader(babelOptions)],
+    },
     resolve: {
       extensions: ['.mjs', '.js', '.jsx', '.json'],
       modules: ['node_modules'].concat(raw.NODE_PATH || []),
