+++ lib/core/src/server/preview/iframe-webpack.config.js
@@ -6,7 +6,8 @@ import CaseSensitivePathsPlugin from 'case-sensitive-paths-webpack-plugin';
 import WatchMissingNodeModulesPlugin from 'react-dev-utils/WatchMissingNodeModulesPlugin';
 import TerserWebpackPlugin from 'terser-webpack-plugin';
 
-import { includePaths, excludePaths, nodeModulesPaths, loadEnv } from '../config/utils';
+import babelLoader from '../common/babel-loader';
+import { nodeModulesPaths, loadEnv } from '../config/utils';
 import { getPreviewHeadHtml, getPreviewBodyHtml } from '../utils/template';
 
 export default ({
@@ -57,17 +58,7 @@ export default ({
     ].filter(Boolean),
     module: {
       rules: [
-        {
-          test: /\.(mjs|jsx?)$/,
-          use: [
-            {
-              loader: 'babel-loader',
-              options: babelOptions,
-            },
-          ],
-          include: includePaths,
-          exclude: excludePaths,
-        },
+        babelLoader(babelOptions),
         {
           test: /\.md$/,
           use: [
