+++ lib/core/src/server/preview/preview-preset.js
@@ -1,19 +1,9 @@
-import loadCustomBabelConfig from '../utils/load-custom-babel-config';
 import loadCustomConfig from '../utils/load-custom-config-file';
-import babelConfig from '../common/babel';
 
 import webpackConfig from './iframe-webpack.config';
 import { createPreviewEntry } from './entries';
 
 export const webpack = async (_, options) => webpackConfig(options);
-
-export const babel = async (_, options) => {
-  const { configDir, presets } = options;
-
-  return loadCustomBabelConfig(configDir, () =>
-    presets.apply('babelDefault', babelConfig(options), options)
-  );
-};
 
 export const entries = async (_, options) => {
   let result = [];
@@ -28,3 +18,5 @@ export const entries = async (_, options) => {
 };
 
 export const config = async (_, options) => loadCustomConfig(options);
+
+export * from '../common/common-preset';
